corrected the Rmidpoint calculation by sorting areas
The Rmidpoint distribution had a suspicious form that looked like an exponential. This was due to an error in the function that calculated Rmidpoint where the cumulative sum of the areas was done on an unsorted array of areas. Now, the distribution looks more like a normal distribution with 3 outliers greater than 800px.